[python] Add the tomli package
For some reason this package is missing, although the python-setuptools-scm depends on it. The apparmor package depends on python-setuptools-scm.
This change can be removed when de packages feed is upstepped to a newer version, since from python 3.11 the tomli package is completely removed as it is part of the default python package.

GitOrigin-RevId: 1a48357fa31b2f32d1eec3377c3aac9b97667e32