libbb/loop: fix existence check for LOOP_CONFIGURE ioctl (CSSP-4204) [fix]
This patch is temporary until we upstream this fix.

The LOOP_CONFIGURE ioctl is supported in 5.8 kernels and up. To have
backwards compatibility there is a config option CONFIG_TRY_LOOP_CONFIGURE
that will check if the ioctl exists and if not fall back to old way of
configuring loop devices.

Normally errno will be set to EINVAL when this ioctl does not exist.
However, when kernel config CONFIG_COMPAT is enabled, then compat_ioctl
is called. In that case -ENOIOCTLCMD is returned by loop device driver
and generic ioctl wrapper will set errno to ENOTTY. Because busybox
does not expect this it will fail to mount loop devices in this case.

This patch fixes the check for the existence of the ioctl LOOP_CONFIGURE
by checking if errno is one of both: EINVAL or ENOTTY.

Issue: CSSP-4204 [PRPL][CI][LCM] LCM tests are failing
GitOrigin-RevId: 064cc9f3a0c8cbf3db66889fac1e6f962ca4213b