Issue: HOP-6290 [miniupnpd] fix incorrect usage of UNPACK_CMD [fix]
The Makefile for miniupnpd-prpl overwrites the UNPACK_CMD to unpack a
subdirectory of the sources to the build directory.

Since autotools is used, OpenWrt provides a set of variables to define
the location of the 'configure' file, but also the path to build the
sources. Therefore, the UNPACK_CMD can be omitted.

GitOrigin-RevId: 7d80648be071214848f61f69ce9e5bf529f3356b