[fstools] Add support of mapped rootfs
When the rootfs is signed and/or ciphered and dmsetup is used, virtual block devices are created and mapped to the physical block device where the rootfs is actually stored. Inside mount_root.c, when volume_find("rootfs_data") is called in start(), the implementation iterates over the registered drivers and eventually when the rootfs is mapped, when calling rootdisk_volume_find(), the virtual device is pointed by rootdev instead of the physical device which should be used.

Hereafter is an example of mapped rootfs (signed and ciphered) stored in a UBI device:
   / # dmsetup ls --tree
   vrootfs_crypt (253:1)
    `-vrootfs_linear (253:0)
       `- (254:0)
   / # cat /sys/class/block/dm-1/dev
   253:1
   / # cat /sys/class/block/dm-0/dev
   253:0
   / # cat /sys/class/block/ubiblock0_5/dev
   254:0
   / # cat /sys/class/ubi/ubi0_5/name
   rootfs
   / # ls -lrt /sys/class/block/dm-1/slaves/
   lrwxrwxrwx    1 root     root             0 Jan  1 00:01 dm-0 -> ../../dm-0
   / # ls -lrt /sys/class/block/dm-1/slaves/dm-0/slaves/
   lrwxrwxrwx    1 root     root             0 Jan  1 00:01 ubiblock0_5 -> ../../ubiblock0_5
   / # ls -lrt /sys/class/block/dm-1/slaves/dm-0/slaves/ubiblock0_5/slaves/

In such a case, the previous implementation failed to detect the rootfs is stored in a UBI and did not mount an overlayfs.
The fix looks for slave devices in /sys/class/block to find the physical block device.

Issue: SBL-856 Use imageconfig to generate PrplOS image
Issue: HOP-3638 Safran is not reboot persistent anymore
GitOrigin-RevId: e2ebae43409d3bc01bcd620a687ac0c36345017c