softhsm2: file storage: use additional encryption layer with KRS
The overall purpose is to make sure that SoftHSM config files (keyring) containing private keys are stored in TMPFS in encrypted form only.
That way an attacker gaining access to the rootfs would not be able to gather the private keys in the clear.

Issue: HOP-6927 Patch SoftHSM to store its keyring in encrypted form only
GitOrigin-RevId: 6fd6ff1ae22fba55546a01c408e2058b0a7020b7