mosquitto: support openssl3 providers for private key
Openssl 3 introduced the providers mechanic to handle specific
implementations and needs, without changing appplications. Among other
things, it allows using a specific backend (pkcs11 for instance) when an
application needs a private key for SSL handshake.

This commit adds support for this feature in mosquitto, by adding support for
openssl3 providers for the private key in mosquitto instances (both for
server and client mode). While the legacy behavior still works, an
application user will be able to ask for a specific private key provider
with provider tags ('provider:some-privkey-specs').

Note that, this commit adds the support as a patch. This patch should be
removed once the change is added in upstream mosquitto

Issue: HOP-7198 - Support openssl3 providers in mosquitto for private key
GitOrigin-RevId: 2c02d5be275638e45e4a1cb0949b39127227da07