libxml2: disable building with libiconv
When switching from autconf to cmake, openwrt started building with libiconv. Since some changes to the core buildsystem were made to accomplish this, we will disable it for now, since we have still an older openwrt environment.
As soon as we switch to newer openwrt version, we can exclude this commit.

GitOrigin-RevId: 026f3a8a2d903018d0fa0c71b9a6abbd80f99426