lighttpd: fix mod_csrf.so installation in staging and rootfs
The mod_csrf.so is compiled when the config option CONFIG_LIGHTTPD_CSRF_PROTECTION is set.
However, we don't install the so in the rootfs.

So let's reuse the BuildPlugin function and create a PACKAGE_lighttpd-mod-csrf.
When this package is enabled, the so file is installed into the rootfs and staging dir.

We keep the CONFIG_LIGHTTPD_CSRF_PROTECTION as well (it simply select PACKAGE_lighttpd-mod-csrf) for two reasons:
- Be backwards compatible
- Make sure the CONFIG_LIGHTTPD_CSRF_PROTECTION cannot be set when no crypto lib is installed (in this case compilation will fail)

Issue: HOP-7328 [lighttpd] mod_csrf.so is not installed in the staging and rooftfs
GitOrigin-RevId: c3c298e70248f7c06df07949cc6ab0ccbdea6b76