apparmor: disable integration with python
After enabling Apparmor, it was seen that python is installed onto the rootfs. This is because Apparmor has a runtime dependency on python by default. This results in the rootfs growing by +- 6MB.

There is however a compiletime option to build Apparmor without python support. Since we don't need the extras that come with python support, we can safely build Apparmor without python integration. This will results in python not being installed into the rootfs and hance saving space.

Issue: HOP-7430 [Integration] Python 3.10 mistakenly embedded in rootfs
GitOrigin-RevId: 9acde08e07c96cf7caf90f1e3ffe2593b8aecf63