lxc: show network config of unprivileged containers (PPW-285)
When an unprivileged container is running with an inheritted network
namespace, then the ip addresses of the container are not shown with
'lxc-ls -f'. This patch will let lxc first enter the net namespace
before entering the user namespace, to prevent permission errors.

Issue: LCMFT-425 network config of unprivileged containers is not shown
Issue: PPW-285 - Network config KO in unprivileged mode

GitOrigin-RevId: fe6df59a1019e3f2d436ffa6a6f76b5bbe0c345d
Signed-off-by: Matthias FRANCK <matthias.franck@softathome.com>