util-linux: allow 32 bit time
Because we are using an older (external) toolchain
for some boxes, it is not possible
to make time_t 64 bits wide, which is what meson.build
tries to do. Allowing 32-bit time tells
meson.build to not give an error during when it fails
to make time_t 64 bits wide.

GitOrigin-RevId: 55d768e46f7dacc0c549120614b57f0a268f6c0b