i2c-tools: change force to recursive at Build/install
When the package was going to be built, the old
package first had to be removed that was in the
PKG_INSTALL_DIR. The RM variable was filled with
"rm -f" and then another -f option was given.

This gave the error that ipkg-install is a directory
and not a file. So the second time -f was supposed
to be -r to have all the files in that directory
removed.

Issue: HOP-8074 add support for openwrt master branch
GitOrigin-RevId: e1b8af4102374d4c65161e3aa8109fcb01c4c687