curl: Use upstream patch for ossl providers support, and fix it
This commit removes a custom patch for openssl3 providers support in
curl library, and adds a patch taken from an upstream commit to the same
effect.
This upstream commit is quite bugged though, so another set of patches
is added to fix the issues.

Issue: HOP-8712 - libcurl crashes with PKCS#11 patch
GitOrigin-RevId: 9c2e8a27ddcd7ae9610521d2cb21c933a92f321f