curl: Use upstream patch for ossl providers support, and fix it
This commit removes a custom patch for openssl3 providers support in
curl library, and adds a patch taken from an upstream commit to the same
effect.
This upstream commit is quite bugged though, so another set of patches
is added to fix the issues.

Issue: HOP-8712 - libcurl crashes with PKCS#11 patch
GitOrigin-RevId: 4384fec9559041dca1cdd9f7a52cbd960aefc616