curl: Add patch to stop overguessing key/cert type
Some code in upstream curl app tries to guess the key or cert type. When
it's not specified on the command line by the user, and it looks like a
pkcs11 URI, the curl app decides its key type is 'ENG' by default.

This behavior goes against openssl3 providers mechanics, which exists to
let the openssl lib try a bunch of types.

This patch removes the code trying to guess the key type. If a user
still wants to use an openssl engine with curl on command line, it's
still possible to do so by explicitely asking for the 'ENG' type.

Issue: HOP-8712 - libcurl crashes with PKCS#11 patch
GitOrigin-RevId: 2021463b50c82cfa35f215706c0520dc8fff2bb5