ubus: Ubusd_proto: remove object availability check in response handler
It is possible to have a situation where a response to a request via
ubus is never received.

The following procedure can be followed to reproduce this issue:
 1) Client A does a lookup for object X (provided by client B) and gets a valid  ID from ubusd
 2) Client A sends a request on object X to ubusd
 3) Ubusd forwards the request to client B
 4) Client B removes object X from ubusd
 5) Client B sends a response to the request from client A to ubusd
 6) Ubusd drops the response because object X is no longer known, causing client A to be waiting forever.

The issue can be avoided by removing the object availability check in
the ubusd response handler. The response from client B will then still
be sent to client A.

Issue: HOP-10003 [Ubusd] No response received when object is removed
GitOrigin-RevId: 163cc14dd0bef44b5c16469ab2b24ca837ebe07b