libubus-obj: Add peer,sequence and object info in status messages
It is possible to have a situation where a response to a request via
ubus is never received.

The issue occurs in the following scenario:
 1) Client A does a lookup for object X (provided by client B) and gets a valid  ID from ubusd
 2) Client A sends a request on object X to ubusd
 3) Client B sends a remove_object message to ubusd (for object X)
 4) Ubusd forwards the request to client B
 5) The request from #4 is queued in client B (at ubus level) because B
    is waiting for a reply for request #3.
 6) Ubusd replies OK to B in response to #3. B removes the ubus object X
    internally as well.
 7) B processes the ubus request #4:
   1) The requested object, X, is no longer known at ubus level.
   2) B sends a status message with status UBUS_STATUS_NOT_FOUND to
      ubusd, without peer, sequence or object info.
 8) Ubusd receives the status message and drops it, since it does not
    know where to send it to.
 9) Client A never receives a response for request #2 and waits until a
    timeout happens.

Adding the peer, sequence and object information in the status message
enables ubusd to forward the UBUS_STATUS_NOT_FOUND message to the
correct client. Client A will now receive an error status instead of
nothing, unblocking the system.

Issue: HOP-10497 [Ubus] Insufficient information in status response messages
GitOrigin-RevId: 277435a1b2f6713524adaf9075e8c2093733a353