libpam: Add patch 0002-change-environment-to-pam_environment.patch to make PrplOS compliant (CSSP-5282)
Problem:
libpam is not PrplOS compliant

Cause:
The module pam_env creates the file `/etc/environment`, which is already
in use in PrplOS and can not be replaced.

Solution:
Add this patch to change the instances of `/etc/environment` to
`/etc/pam_environment`. This file name can be used without any issue.

Issue: CSSP-5282 - "[WIFI7] Incompatibility Issue with fscrypt Binary on PrplOS"
GitOrigin-RevId: fe670116003d2ab5b93203dc174d3811d6710aa5