libxslt: import 1.1.42
Fetched from: https://github.com/openwrt/packages/blob/master/libs/libxslt/
libxslt has a dependency on libxml2 which was recently upstepped for CVE reasons.
There are now compile issues with libxslt, so we also upstep libxslt to a newer version to work well with the new libxml2.

Issue: HOP-9878 [CVE] libxml2 as identified vulnerabilities and shoulld be upsteped to min 2.13.8 or 2.14.2 or more recent
GitOrigin-RevId: f3dfd0bdf19391eedeac442b03eca2d22b58b954