p11-kit client relies on reading environment variables.
Access to environment variables will fail for processes that changed userid, groupid or capabilities. This is the case for plugins that have configurations to not execute as root.

To have this type of plugins use pkcs11 uris, we need to have some way to access the information set in some environment variables. This patch adds a build option to allow reading of the needed environment variables, it can be set to true to remove all access again.

original commit: 8ad9ede5

Issue: SBL-1487 - Implement QSEE PKSC#11 library
GitOrigin-RevId: 8afc65f4c09120244ddd582df52bc69cb5bad199