mstpd: ignore vlan tagged BPDUs
Due to the fact that the kernel strips VLAN tags, those tags are
not part of the data received by the raw packet socket that captures
BPDUs. This can cause problems in the following situation:

Suppose that a BPDU with VLAN tag 10 is received on the physical
interface eth2. If the interface eth2.10 does not exist, the BPDU
will be processed by mstpd as if it were received untagged on eth2.
This can lead to STP not converging due to conflicting BPDUs.

To solve this, following adaptations were made to include VLAN information
to the packets received by mstpd:
    - set the packet socket option PACKET_AUXDATA to request auxiliary
      data which can contain VLAN information.
    - use the TPACKET_V2 packet version.
    - change the socket protocol to ETH_P_ALL.
    - use recvmsg instead of recvfrom.

When a BPDU has been received that has a VLAN tag, it is silently dropped.
BPDUs which originally had a VLAN tag but are received on a VLAN tagged
interface will be processed normally, as the VLAN tag is not part of the
auxiliary information.

GitOrigin-RevId: f8da1034000204ab8fbeb4a8bb8199ff7d617271