sah: set PKG_MIRROR_HASH to "skip" for packages cloning from git
The PKG_MIRROR_HASH is very dependant on the openwrt version (versions of tar, compression format, ... differs).
As a result, this hash can only be calculated for a very specific openwrt version.

Since we want to be able to use our feeds on more openwrt versions in parallel, we cannot use this feature.

In openwrt <= 24.10 this wrong hash leads to a recompilation of the package every time "make" is called.
In openwrt >24.10 this leads to a full stop of the build.

GitOrigin-RevId: 98c1a83e3dee4c28d1b5abc1cb50693429ccedef