xcrypt: import libxcrypt and libcrypt-compat
Sudo was bumped to a newer version, taken from the openwrt master branch.
With this upsted, a dependency on libxcrypt-compat was introduced.

This package is however missing in our current version of the buildsystem used.
So let's add the package here for now.
With a new import of openwrt buildsystem the package will become availble and we can remove this import.

Imported from openwrt master branch at 5f99bcdbecbbf816c0fcddd7fd32f5a9d511791b

GitOrigin-RevId: 6591dead76d3b39a9fc4bbf117c8ce6603e8ce6e