libcrypt-compat: build dummy package when building with glibc<2.39
This library is included in glibc<2.39 by default. Building this would lead to runtime problems.
Making this package a conditional dependency for all openwrt packages that depend on it would mean a lot of forks in feed_opensource.

In stead, we opt for a solution where a dummy package for libcrypt-compat is installed when building with glibc<2.39.
This way openwrt compilation is not broken, but nothing will be installed in practice when this dummy package is installed.

Issue: HOP-11350
GitOrigin-RevId: 4fa191ff3619f06a334f572863462a8efd316375