radvd: allow spoofing of RA source address
radvd allows to specify the RA source address using the parameter
AdvRASrcAddress. A limitation is that the specified address must
belong to the interface on which the RA is sent. This limitation
makes sense in most cases, but sometimes you want to use a non-local
address. When RAs are sent from a repeater, it's useful to spoof the
RA source address with the LLA of the gateway instead so that IPv6
traffic from clients connected to the repeater is bridged to the
gateway. Therefore, this patch removes the check to see whether
the configured AdvRASrcAddress belongs to the interface.

Note: for the spoofing to work, /proc/sys/net/ipv6/ip_nonlocal_bind
must be set to 1.

Issue: HOP-8859 ipv6 guest connectivity on repeater is broken.
GitOrigin-RevId: 890d2b30b41c168f52ec0b836b74c07547eaff2d