lvm2: Enable install with write permission (CSSP-5831)
Problem:
The debug symbols from package libdevmapper.so can not be stripped.

Cause:
libdevmapper.so is installed without write permissions, even for root.
Because of this, sstrip does not have permission to write the stripped
package.

Solution:
Enable the `write_install` mode. With this, the file is installed with
write permissions for root.

Issue: CSSP-5831 - "libdevmapper unstripped with debug information"
GitOrigin-RevId: 2a465ccde2e54892e7661f9e2b650423da5cbb1b